Update fabric_doc_open_revs to handle purges
We need to account for the possibility that a document is opened while a
purge is still propogating between shards. This means we need to inform
the read-repair algorithms that a difference in revisions may be due to
a purge request in progress. If we don't do this its possible that a
read-repair may race the purge request and effectively undo the purge.